modified the timeslice - if didn't make sense to me as it was.
RV is supposed to be a set of possible values, but the existing
implementation used a Stream of identical values which made the discount
calculation easier.

My implementation uses a single value for each timeslice which I think
makes sense for time. I needed to modify the discount calculation but
that needs to be more sophisticated anyway.